[remove] BigInteger 저장문제로 압수
+System.Numerics에 있대서 스크립트도 죽임

+++ Assets/02.Scripts/Data/UserData.cs
@@ -7,8 +7,8 @@ public class UserData
     public string userId;
     public string userName = "Name한글";
 
-    public BigInteger gold = 1000;
-    public BigInteger diamond = 100;
+    public int gold = 1000;
+    public int diamond = 100;
     public int seed = 10;
     public int food = 10;
     //public int key;
